feat: Validates duplicate thresholds
